# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yoko API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/words": {
            "get": {
                "tags": [
                    "Word"
                ],
                "summary": "Get all words (Route only for admins)",
                "operationId": "a51c08df4002d393e38001de64aee284",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query for searching words by english meaning",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "words": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "word": {
                                                                "type": "string",
                                                                "example": "this"
                                                            },
                                                            "translationUk": {
                                                                "type": "string",
                                                                "example": "цей"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/words/{id}": {
            "get": {
                "tags": [
                    "Word"
                ],
                "summary": "Get word by id (Route only for admins)",
                "operationId": "b5e46ed00e384318db63227e4b62d982",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of word",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "word": {
                                                    "type": "string",
                                                    "example": "this"
                                                },
                                                "translationUk": {
                                                    "type": "string",
                                                    "example": "цей"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Word"
                ],
                "summary": "Update word",
                "operationId": "b19121d2ee40a1bf9cc3171d645e1ec6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of word",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "word": {
                                                "type": "string",
                                                "example": "dog"
                                            },
                                            "translationUk": {
                                                "type": "string",
                                                "example": "собака"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Request successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "word": {
                                                    "type": "string",
                                                    "example": "dog"
                                                },
                                                "translationUk": {
                                                    "type": "string",
                                                    "example": "собака"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}